backend/requirements.txt
Django==6.0.2
djangorestframework==3.14.0
django-cors-headers==4.0.0
Pillow>=9.0.0
